refactor(settings): 抽取 SettingRow 为共享组件并同步分组设置页展示密度
- SettingRow 从设置页移至 components/config，替换分组设置页运行时区块里的
  RuntimeOverrideRow，行为对齐：默认态收敛徽标/按钮、覆盖态才展开控件与说明。
- RuntimeOverrideRow 无剩余引用，一并删除。
- 顺带修正分组页 inject_usage_options 折叠态展示 help 文案而非当前值的问题，
  改为展示已启用/已禁用，说明文字移入 help 提示。
- 请求亲和沿用原有 SegmentedControl 三态控件，未纳入本次迁移。
