+++ web/src/features/settings/ReliabilitySettingsSection.vue
@@ -9,7 +9,7 @@ import type {
 import AppTextInput from '@/components/ui/AppTextInput.vue'
 import CompactFieldError from '@/components/ui/CompactFieldError.vue'
 
-import SettingRow from './SettingRow.vue'
+import SettingRow from '@/components/config/SettingRow.vue'
 import {
   createSettingsDraft,
   isValidNonNegativeInteger,
